feat: custom derivation path picker for ledger polkadot legacy
