+++ apps/extension/src/ui/domains/Account/LedgerPolkadotLegacyAccountPicker/types.ts
@@ -1,0 +1,9 @@
+import { LedgerAccountDefSubstrate } from "../AccountAdd/AccountAddLedger/context"
+import { DerivedAccountBase } from "../DerivedAccountPickerBase"
+
+export type LedgerPolkadotLegacyAccountPickerProps = {
+  chainId: string
+  onChange?: (accounts: LedgerAccountDefSubstrate[]) => void
+}
+
+export type LedgerPolkadotAccountPickerDef = DerivedAccountBase & LedgerAccountDefSubstrate
